+++ changelog.md
@@ -2,6 +2,7 @@
 
 - Added the command `/ess convert PlauerVaultX`, allows converting the database from PlayerVaultX to zEssentials.
 - Added the command `/ess convert Sunlight`, allows converting the database from Sunlight to zEssentials.
+- Added the command `/ess convert CoinsEngine`, allows converting the database from CoinsEngine to zEssentials.
 - Items in player vault will be displayed as a single item.
 
 # 1.0.0.3